Temporarily fix React hook rules and TS errors js_only
